Moving events and form validators to seperate files.
